Updated float in kick_factor

+++ include/filemanager.hpp
@@ -94,7 +94,7 @@ class SimTasker {
 
 	  bool is_infinite;
 	  int time_allowed;
-	  int kick_factor;
+	  float kick_factor;
 
    private:
 	  nlohmann::json allData;
